requirements.txt
# === AI Pipeline ===
whisperx
faster-whisper
groq
sentence-transformers
pyspellchecker
librosa
soundfile
numpy
langdetect

# === Backend (Full Stack - NEW) ===
fastapi
uvicorn[standard]
aiosqlite
python-multipart
aiofiles
python-dotenv

# === Video Editor ===
av>=12.0.0
Pillow>=10.0.0
